TODO: footer slots. Removes FlowItem. Enhances Panel to handle FlowItem concerns. Updates Flow to use Panel. (#891)
