--- src/components/calcite-flow-item/calcite-flow-item.scss
@@ -1,42 +1,0 @@
-:host {
-  @extend %component-host;
-  background-color: var(--calcite-ui-background);
-  display: flex;
-  height: 100%;
-  width: 100%;
-}
-
-calcite-panel {
-  width: 100%;
-  height: 100%;
-}
-
-.header-leading-actions {
-  display: flex;
-}
-
-.menu-button {
-  align-self: stretch;
-  flex: 0 1 auto;
-  height: 100%;
-  position: relative;
-}
-
-.header-trailing-actions,
-.header-single-action-container {
-  display: flex;
-}
-
-.menu {
-  min-width: var(--calcite-menu-min-width);
-  flex-flow: column nowrap;
-}
-
-.footer-actions {
-  display: flex;
-  width: 100%;
-}
-
-.fab-container {
-  display: inline-block;
-}
