--- src/components/calcite-flow-item/resources.ts
@@ -1,32 +1,0 @@
-export const CSS = {
-  header: "header-content",
-  backButton: "back-button",
-  footerActions: "footer-actions",
-  headerTrailingActions: "header-trailing-actions",
-  headerLeadingActions: "header-leading-actions",
-  headerSingleActionContainer: "header-single-action-container",
-  menuContainer: "menu-container",
-  menuButton: "menu-button",
-  menu: "menu",
-  menuOpen: "menu--open",
-  fabContainer: "fab-container"
-};
-
-export const SLOTS = {
-  headerLeadingActions: "header-leading-actions",
-  headerTrailingActions: "header-trailing-actions",
-  fab: "fab",
-  footerActions: "footer-actions"
-};
-
-export const ICONS = {
-  menu: "ellipsis",
-  backLeft: "chevron-left",
-  backRight: "chevron-right"
-};
-
-export const TEXT = {
-  back: "Back",
-  open: "Open",
-  close: "Close"
-};
