--- src/demos/flow-item/basic.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=5.0" />
-
-    <title>Calcite Components: calcite-flow-item</title>
-
-    <script src="../_assets/head.js"></script>
-    <script src="../_assets/toggles.js"></script>
-  </head>
-
-  <body>
-    <main>
-      <section class="example-container">
-        <h1>calcite-flow-item</h1>
-
-        <h2>Flow Item</h2>
-        <div class="flow-container">
-          <calcite-flow-item heading="What are the most popular commute alternatives?" summary="I don't know.">
-            <calcite-action slot="menu-actions" text="Save" text-enabled></calcite-action>
-            <calcite-action slot="menu-actions" text="Do It" text-enabled></calcite-action>
-            <calcite-action slot="menu-actions" text="Load" text-enabled></calcite-action>
-            <calcite-action slot="menu-actions" text="Cancel" text-enabled></calcite-action>
-            <button slot="footer-actions">Example</button>
-            <button slot="footer-actions">Actions</button>
-            <img src="https://via.placeholder.com/300x200" alt="placeholder" />
-          </calcite-flow-item>
-        </div>
-
-        <h2>Scrollable Flow Item</h2>
-        <div class="flow-container" style="height: 500px;" id="flow-item">
-          <calcite-flow-item heading="What are the most popular commute alternatives?" summary="I don't know.">
-            <calcite-action slot="menu-actions" text="Save" text-enabled></calcite-action>
-            <calcite-action slot="menu-actions" text="Do It" text-enabled></calcite-action>
-            <calcite-action slot="menu-actions" text="Load" text-enabled></calcite-action>
-            <calcite-action slot="menu-actions" text="Cancel" text-enabled></calcite-action>
-            <button slot="footer-actions">Example</button>
-            <button slot="footer-actions">Actions</button>
-            <img src="https://via.placeholder.com/300x200" alt="placeholder" />
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-            <p>Slotted content</p>
-          </calcite-flow-item>
-        </div>
-        <script>
-          const flowItem = document.getElementById("flow-item");
-          flowItem.addEventListener("calciteFlowItemScroll", function (event) {
-            console.log("scrolled", event);
-          });
-        </script>
-      </section>
-    </main>
-  </body>
-</html>
